Move Ulam spiral generator to its own class
