+++ Content.Server/Tabletop/TabletopSystem.Map.cs
@@ -1,5 +1,6 @@
 using System.Numerics;
 using Content.Shared.GameTicking;
+using Content.Shared.Maths;
 using Robust.Shared.Map;
 using Robust.Shared.Map.Components;
 
@@ -37,7 +38,7 @@ namespace Content.Server.Tabletop
         /// <returns></returns>
         private Vector2 GetNextTabletopPosition()
         {
-            return UlamSpiral(++_tabletops) * TabletopSeparation;
+            return UlamSpiral.Point(++_tabletops) * TabletopSeparation;
         }
 
         /// <summary>
@@ -57,34 +58,6 @@ namespace Content.Server.Tabletop
             // Lighting is always disabled in tabletop world.
             mapComp.LightingEnabled = false;
             Dirty(mapUid, mapComp);
-        }
-
-        /// <summary>
-        ///     Algorithm for mapping scalars to 2D positions in the same pattern as an Ulam Spiral.
-        /// </summary>
-        /// <param name="n">Scalar to map to a 2D position. Must be greater than or equal to 1.</param>
-        /// <returns>The mapped 2D position for the scalar.</returns>
-        private Vector2i UlamSpiral(int n)
-        {
-            var k = (int)MathF.Ceiling((MathF.Sqrt(n) - 1) / 2);
-            var t = 2 * k + 1;
-            var m = (int)MathF.Pow(t, 2);
-            t--;
-
-            if (n >= m - t)
-                return new Vector2i(k - (m - n), -k);
-
-            m -= t;
-
-            if (n >= m - t)
-                return new Vector2i(-k, -k + (m - n));
-
-            m -= t;
-
-            if (n >= m - t)
-                return new Vector2i(-k + (m - n), k);
-
-            return new Vector2i(k, k - (m - n - t));
         }
 
         private void OnRoundRestart(RoundRestartCleanupEvent _)
